Revert the namespace refactor
It just adds too much noise to the PR.
Perhaps as a separate future PR it could be introduced again.
