+++ spec/presenters/vlo_history_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Timeline::VloHistory do
+RSpec.describe VloHistory do
   let(:event) { 'create' }
   let(:version) { PaperTrail::Version.new(event:) }
   let(:presenter) { described_class.new(version) }
