+++ spec/services/allocation_service_spec.rb
@@ -148,58 +148,58 @@ describe AllocationService do
 
     let(:case_histories) do
       [
-        double(Timeline::CaseHistory, created_at: '2020-04-09T02:01', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2020-04-09T13:27', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2020-04-09T13:28', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2020-08-19T11:10', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2020-08-19T11:28', primary_pom_name: 'DICKS:', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2020-08-19T11:29', primary_pom_name: 'JENNINGS', primary_pom_email: nil, secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-03-10T10:52', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-04-28T10:23', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-06-02T10:59', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-06-02T10:59', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'ADEOYE', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-06-11T13:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'ADEOYE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-06-21T11:03', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-06-21T11:05', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-03T07:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-03T07:53', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-03T08:29', primary_pom_name: 'ADEOYE', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-03T09:27', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-03T11:38', primary_pom_name: 'ADEOYE', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-04T09:56', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-04T12:37', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-04T13:40', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-10T11:09', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-11T14:26', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-11T14:26', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-12T09:24', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-08-13T15:41', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-12-29T15:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2021-12-29T15:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'BATLINER:ANIE', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'BATLINER:ANIE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'BATLINER:ANIE', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-01T15:10', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'BATLINER:ANIE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-01T15:10', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-01T15:11', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-08T14:51', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-09T14:06', primary_pom_name: 'ROOKER:', primary_pom_email: 'rooker@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:51', primary_pom_name: 'ROOKER:', primary_pom_email: 'rooker@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:55', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T08:58', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-03-31T11:48', primary_pom_name: 'BATLINER:', primary_pom_email: 'batliner@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-04-22T15:08', primary_pom_name: 'POM:', primary_pom_email: 'pom@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-04-25T12:54', primary_pom_name: 'POM:', primary_pom_email: 'pom@x.com', secondary_pom_name: 'ROOKER:', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-10-26T11:41', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: 'ROOKER:', event: 'reallocate_primary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-11-21T17:59', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2022-11-21T17:59', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: 'QAZI, ASFAND', event: 'allocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2023-02-08T12:12', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
-        double(Timeline::CaseHistory, created_at: '2023-02-16T12:13', primary_pom_name: 'DUCKETT', primary_pom_email: nil, secondary_pom_name: nil, event: 'reallocate_primary_pom')
+        double(CaseHistory, created_at: '2020-04-09T02:01', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
+        double(CaseHistory, created_at: '2020-04-09T13:27', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2020-04-09T13:28', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2020-08-19T11:10', primary_pom_name: 'BAYRAM:', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2020-08-19T11:28', primary_pom_name: 'DICKS:', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2020-08-19T11:29', primary_pom_name: 'JENNINGS', primary_pom_email: nil, secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-03-10T10:52', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-04-28T10:23', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'MCCANN, KATHERINE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-06-02T10:59', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2021-06-02T10:59', primary_pom_name: 'DICKS', primary_pom_email: 'dicks@x.com', secondary_pom_name: 'ADEOYE', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2021-06-11T13:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'ADEOYE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-06-21T11:03', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2021-06-21T11:05', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2021-08-03T07:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-03T07:53', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-03T08:29', primary_pom_name: 'ADEOYE', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-03T09:27', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-03T11:38', primary_pom_name: 'ADEOYE', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-04T09:56', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-04T12:37', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-04T13:40', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-10T11:09', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-11T14:26', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-11T14:26', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-12T09:24', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-08-13T15:41', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'JARA DUNCAN, LAURA', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2021-12-29T15:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2021-12-29T15:46', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'BATLINER:ANIE', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'BATLINER:ANIE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-01-27T08:20', primary_pom_name: 'CUFFLIN', primary_pom_email: nil, secondary_pom_name: 'BATLINER:ANIE', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-03-01T15:10', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'BATLINER:ANIE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-01T15:10', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-03-01T15:11', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-03-08T14:51', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-09T14:06', primary_pom_name: 'ROOKER:', primary_pom_email: 'rooker@x.com', secondary_pom_name: 'CASTIGLIONE', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:51', primary_pom_name: 'ROOKER:', primary_pom_email: 'rooker@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:55', primary_pom_name: 'BAYRAM', primary_pom_email: 'bayram@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:56', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T08:58', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-03-31T11:48', primary_pom_name: 'BATLINER:', primary_pom_email: 'batliner@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-04-22T15:08', primary_pom_name: 'POM:', primary_pom_email: 'pom@x.com', secondary_pom_name: nil, event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-04-25T12:54', primary_pom_name: 'POM:', primary_pom_email: 'pom@x.com', secondary_pom_name: 'ROOKER:', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-10-26T11:41', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: 'ROOKER:', event: 'reallocate_primary_pom'),
+        double(CaseHistory, created_at: '2022-11-21T17:59', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2022-11-21T17:59', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: 'QAZI, ASFAND', event: 'allocate_secondary_pom'),
+        double(CaseHistory, created_at: '2023-02-08T12:12', primary_pom_name: 'ADAMS:', primary_pom_email: 'adams@x.com', secondary_pom_name: nil, event: 'deallocate_secondary_pom'),
+        double(CaseHistory, created_at: '2023-02-16T12:13', primary_pom_name: 'DUCKETT', primary_pom_email: nil, secondary_pom_name: nil, event: 'reallocate_primary_pom')
       ]
     end
 
@@ -235,8 +235,8 @@ describe AllocationService do
     context 'when the current primary POM is deallocated without a replacement' do
       let(:case_histories) do
         [
-          double(Timeline::CaseHistory, created_at: '2024-01-10T09:00', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
-          double(Timeline::CaseHistory, created_at: '2024-02-11T10:30', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_primary_pom')
+          double(CaseHistory, created_at: '2024-01-10T09:00', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
+          double(CaseHistory, created_at: '2024-02-11T10:30', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_primary_pom')
         ]
       end
 
@@ -250,10 +250,10 @@ describe AllocationService do
     context 'when the same POM is allocated again after a deallocation' do
       let(:case_histories) do
         [
-          double(Timeline::CaseHistory, created_at: '2024-01-10T09:00', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
-          double(Timeline::CaseHistory, created_at: '2024-02-11T10:30', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_primary_pom'),
-          double(Timeline::CaseHistory, created_at: '2024-03-12T11:45', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
-          double(Timeline::CaseHistory, created_at: '2024-04-13T12:00', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_released_offender')
+          double(CaseHistory, created_at: '2024-01-10T09:00', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
+          double(CaseHistory, created_at: '2024-02-11T10:30', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_primary_pom'),
+          double(CaseHistory, created_at: '2024-03-12T11:45', primary_pom_name: 'SMITH, JANE', primary_pom_email: 'smith@example.com', secondary_pom_name: nil, event: 'allocate_primary_pom'),
+          double(CaseHistory, created_at: '2024-04-13T12:00', primary_pom_name: nil, secondary_pom_name: nil, event: 'deallocate_released_offender')
         ]
       end
