+++ spec/support/shared_examples/paper_trail_history_presenter_examples.rb
